+++ src/components/index.js
@@ -236,4 +236,6 @@ export {
 export {
   EuiText,
   EuiTextColor,
+  EuiTextAlign,
+  EuiTextFontFamily,
 } from './text';