+++ src/components/text/text_font_family.test.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { render } from 'enzyme';
+import { requiredProps } from '../../test';
+
+import { EuiTextFontFamily } from './text_font_family';
+
+describe('EuiTextFontFamily', () => {
+  test('is rendered', () => {
+    const component = render(
+      <EuiTextFontFamily fontFamily="code" {...requiredProps} />
+    );
+
+    expect(component)
+      .toMatchSnapshot();
+  });
+});