+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - Updated `<EuiErrorBoundary>` to preserve newlines in error. [(#238)[https://github.com/elastic/eui/pull/238]]
 - Added more icons and fixed a few for dark mode [(#228)](https://github.com/elastic/eui/pull/228)
 - Added `EuiFlyout` component. [(#227)](https://github.com/elastic/eui/pull/227)
+- Added `EuiTextFontFamily` and `EuiTextAlign` component. [(#249)](https://github.com/elastic/eui/pull/249)
 
 **Breaking changes**